+++ header.php
@@ -42,6 +42,11 @@
 						
 						?>
 					</a>
+					<?php if ( is_active_sidebar( 'header' ) ): ?>
+					<div class="widget-area col-xs-8" role="complementary">
+						<?php dynamic_sidebar( 'header' ); ?>
+					</div>
+					<?php endif; ?>
 					<?php if ( ecologie_get_theme_mod_or_default( 'site_description_primary_menu_on' ) ): ?><p class="navbar-text hidden-md hidden-sm hidden-xs"><?php echo ecologie_shortened_site_description(); ?></p><?php endif; ?>
 				</div>
 				<?php if ( production_mode_disabled() ): ?><span class="pull-right navbar-text"> | Dev mode</span><?php endif; ?>
@@ -56,11 +61,6 @@
 							) );
 						}
 						
-						if ( is_active_sidebar( 'header' ) ): ?>
-							<div class="widget-area pull-right" role="complementary">
-								<?php dynamic_sidebar( 'header' ); ?>
-							</div>
-						<?php endif;
 					?>
 				</div>
 			</div>